Percobaan 1 JS 8

+++ UTS/Soal3/Main.java
@@ -6,13 +6,13 @@ public class Main {
 
         Kendaraan mobil = new Mobil("N 1234 AB", "Toyota", 2020);
         mobil.tampilkanData();
-        System.out.println("Biaya Servis: Rp" + (int)mobil.hitungBiayaServis());
-        System.out.println("Biaya Servis + Cuci: Rp" + (int)mobil.hitungBiayaServis(true));
+        System.out.println("Biaya Servis: Rp" + mobil.hitungBiayaServis());
+        System.out.println("Biaya Servis + Cuci: Rp" + mobil.hitungBiayaServis(true));
         System.out.println();
 
         Kendaraan motor = new Motor("N 5678 XY", "Honda", 2022);
         motor.tampilkanData();
-        System.out.println("Biaya Servis: Rp" + (int)motor.hitungBiayaServis());
-        System.out.println("Biaya Servis + Cuci: Rp" + (int)motor.hitungBiayaServis(true));
+        System.out.println("Biaya Servis: Rp" + motor.hitungBiayaServis());
+        System.out.println("Biaya Servis + Cuci: Rp" + motor.hitungBiayaServis(true));
     }
 }
